refactor: 프론트 WondowSize를 util로 이동
